app: extract ShortcutHandler to a separate class
